+++ pennylane_lightning/lightning_base/_serialize.py
@@ -79,7 +79,7 @@ class QuantumScriptSerializer:
                 ) from exception
         elif device_name == "lightning.kokkos":
             try:
-                import pennylane_lightning.lightning_kokkos_ops as lightning_ops
+                import pennylane_lightning.lightning_kokkos_nb as lightning_ops
             except ImportError as exception:
                 raise ImportError(
                     f"Pre-compiled binaries for {device_name} are not available."